Spit mc-java extension tests
... trying to save on creating temp. projects.
